+++ server/src/main/java/CryptOptima/server/security/dto/LoginDto.java
@@ -1,8 +1,14 @@
 package CryptOptima.server.security.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 public class LoginDto {
     private String accountId;
     private String password;
